Fix truncation of paths for finder provider
The finder provider tries to shorten the path according to the language
servers root directory or the users home directory. This was not always
working correctly. The issue behind this was that the used matching and
substituation functions use pattern(magic) per default. So the used
path variables are not handled as plain strings, but get interpreted.
For the matching part, there is a parameter to disable pattern matching
and do plain version. But for the used substitution this does not work
that easy. Escaping the string is too cumbersome. The way more simple
approach is just to cut off the beginning of the long path by its
origin. Since we do the matching specifically from index 1, this will
always work as expected. Therefore it is the more simple and straight
forward solution.