Open all folds for floating preview windows (#31)
Having folds in a preview of a definition for example is quite annoying.
You can't read everything, but also not really open them nicely (besides
you don't want to go through this hassle). Furthermore do the dimensions
of the window not fit anymore as the folds reduce the size of the code
part that relates to the definition.