Windows lsp_finder fixes (#87)
* Windows lsp_finder fixes

vim.loop.os_uname().sysname appears to return "Windows_NT", so the
in_windows check was changed to check for a substring instead of
straight equality. This was also migrated to a function to perform a
similar check in provi.ers.lua.

os.getenv("HOME") returns nil on Windows. os.getenv("USERPROFILE") is
appropriate on Windows.

vim.uri_to_fname() returns paths with lowercase drive letters on Windows
for some reason. Before the paths are shortened for the lsp_finder
results, a check for Windows was added, along with a substitution of the
drive letter for the appropriate uppercase version.

* removed trailing whitespace

* More fixes for Windows