fix: allow user to pass data into codeaction API
Previously, all data that was passed as arguments into one of these
functions would be discarded, since these functions are defined to take
no arguments.

This makes it so that these outer API functions pass all arguments
directly to the inner functions. This allows the user to for example
press "2" and actually have "2" transmitted to the inner function,
instead of being lost.

This also may likely prevent a whole host of other potential bugs.