open finder reference in the calling window
This is a small convenience improvement. Currently, when the finder uses
a normal layout with sp_global, the reference is always open
('toggle_or_open') in the leftmost window. This is quite inconvenient
when working with multiple vertical splits. Imagine that you search for
reference in the right split, and then the reference is open in the left
split, which has a completely different context.

When the finder is open, we memorize the current window id. When the
reference is opened, we use that window id instead of 0 (which by this
time is already the leftmost).

For float layout, or normal with sp_global=false, there is no change in
behavior.